lint and organize types in UnidocBuild
